requirements.txt
astroid==2.0.4
autopep8==1.4
Django==2.1
django-debug-toolbar==1.10.1
djangorestframework==3.8.2
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
Pillow==5.1.0
pycodestyle==2.4.0
pylint==2.1.1
pytz==2018.5
six==1.11.0
sqlparse==0.2.4
typed-ast==1.1.0
wrapt==1.10.11
